Compute values base on property instead
