+++ components/style/src/value_processing.rs
@@ -19,8 +19,7 @@ use super::expand::prelude::*;
 
 // computes
 use super::computes::color::compute_color;
-use super::computes::percentage::compute_percentage;
-use super::computes::length::compute_length;
+use super::computes::font_size::compute_font_size;
 
 type DeclaredValuesMap = HashMap<Property, Vec<PropertyDeclaration>>;
 
@@ -150,10 +149,9 @@ pub fn apply_styles(node: &NodeRef, rules: &[ContextualRule]) -> Properties {
 
 /// Resolve specified values to computed values
 pub fn compute(property: &Property, value: &Value, context: &mut ComputeContext) -> ValueRef {
-    match value {
-        Value::Color(_) => compute_color(value, property, context),
-        Value::Percentage(_) if matches!(property, Property::FontSize) => compute_percentage(value, property, context),
-        Value::Length(_) if matches!(property, Property::FontSize) => compute_length(value, property, context),
+    match property {
+        Property::Color => compute_color(value, context),
+        Property::FontSize => compute_font_size(value, context),
         _ => {
             if !context.style_cache.contains(value) {
                 context.style_cache.insert(ValueRef::new(value.clone()));
